Fix the dev-integration build
It turns our there is an issue with our bsltestutil facility that
does not show up on modern compilers that happen to implement the
C++11 support for instantiating templates on types with internal
linkage.  In this case, we have an un-named enumeration type used
to signify some alignment property.  As BDE used un-named enum
types rather than static const integral data members to indicate
compile time constants, this may be a lurking issue that will
crop up again.  The immediate concern is resolved by casting a
single enum value to 'int' inside the 'P(X)' invocation, and now
the whole of the blsma facility converted to bsltestutil should
build and run correctly in the nightly tests.

+++ groups/bsl/bslma/bslma_infrequentdeleteblocklist.t.cpp
@@ -391,7 +391,7 @@ int main(int argc, char *argv[])
         const int BLKSZ = sizeof(InfrequentDeleteBlock)
                                                        - U::BSLS_MAX_ALIGNMENT;
 
-        if (veryVerbose) { T_ P_(BLKSZ); P(U::BSLS_MAX_ALIGNMENT); }
+        if (veryVerbose) { T_ P_(BLKSZ); P((int)U::BSLS_MAX_ALIGNMENT); }
 
         struct {
             int d_line;      // line number
